+++ src/bin/hnefatafl-client.rs
@@ -1185,12 +1185,15 @@ impl Client {
                     }
                 }
 
+                let (attacker_pieces_taken, defender_pieces_taken) = game.board.pieces_taken();
                 let user_area_ = column![
                     row![
                         text(self.time_attacker.fmt_shorthand()).size(40).center(),
                         text("⚔").shaping(text::Shaping::Advanced).size(40).center(),
                         column![
-                            text(self.attacker.to_string()).center(),
+                            text(format!("{} ♙ {defender_pieces_taken}", self.attacker))
+                                .shaping(text::Shaping::Advanced)
+                                .center(),
                             text(attacker_rating.to_string()).center(),
                         ]
                     ]
@@ -1199,7 +1202,9 @@ impl Client {
                         text(self.time_defender.fmt_shorthand()).size(40).center(),
                         text("🛡").shaping(text::Shaping::Advanced).size(40).center(),
                         column![
-                            text(self.defender.to_string()).center(),
+                            text(format!("{} ♟ {attacker_pieces_taken}", self.defender))
+                                .shaping(text::Shaping::Advanced)
+                                .center(),
                             text(defender_rating.to_string()).center(),
                         ]
                     ]